# packages/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: NEBULA Control API
  version: "1.0.0-phase2"
  summary: Model registry, credentials and deployment lifecycle.
  description: |
    The control plane is the only writer of desired state. Everything here is
    org-scoped by the caller's credential: `org_id` is never a parameter, in a
    path, a query or a body.

    **Deployments are reconciled intent.** A successful `POST /v1/deployments`
    returns `202 Accepted` and records desired state in PostgreSQL. The Kubernetes
    controller that acts on it arrives in Phase 5, so in this build
    `status.reconciled` is always `false` and `status.observed_generation` is
    always `0`. Responses that describe unreconciled state carry a `note` saying
    so, rather than leaving a caller to infer that work is underway.
  license:
    name: Apache-2.0
  contact:
    name: NEBULA
    url: https://github.com/adityasatwar321/nebula
servers:
  - url: /
    description: The service itself, reached through the gateway in a deployed cluster.
security:
  - apiKey: []
...
paths:
...
  /v1/users:
    get:
      tags: [identity]
      operationId: listUsers
      summary: List users.
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: A page of users.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [identity]
      operationId: createUser
      summary: Create a user.
      description: |
        Only an owner may create an owner. Without that rule the role hierarchy is
        decorative, because any admin could mint an owner and take the
        organization.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUser" }
      responses:
        "201":
          description: The created user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/users/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [identity]
      operationId: getUser
      summary: Fetch a user.
      responses:
        "200":
          description: The user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [identity]
      operationId: updateUser
      summary: Change a user's role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: "#/components/schemas/UserRole" }
      responses:
        "200":
          description: The updated user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [identity]
      operationId: deleteUser
      summary: Remove a user.
      description: |
        A soft delete. Refused for the last owner: an organization with no owner
        cannot be administered or billed.
      responses:
        "204": { description: Removed. }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/api-keys:
    get:
      tags: [identity]
      operationId: listAPIKeys
      summary: List API keys, including revoked ones.
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: A page of keys. The secret is never included.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/APIKey" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [identity]
      operationId: createAPIKey
      summary: Mint an API key.
      description: |
        The plaintext key is returned in this response and never again: the server
        stores an HMAC-SHA256 with a server-side pepper and cannot reproduce the
        secret.

        A caller cannot grant a scope its own credential does not carry. Otherwise
        `admin` would be a privilege-escalation primitive and scoping would be
        advisory.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAPIKey" }
      responses:
        "201":
          description: The created key, with its one-time secret.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedAPIKey" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/api-keys/{key_id}:
    parameters:
      - name: key_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [identity]
      operationId: revokeAPIKey
      summary: Revoke an API key.
      description: |
        Idempotent. The revoking process stops honouring the key immediately; other
        control-plane replicas honour it for at most `auth.key_cache_ttl`, because
        revocation broadcast arrives in Phase 4. That window is why the TTL default
        is 30 seconds.
      responses:
        "204": { description: Revoked. }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/audit-logs:
    get:
      tags: [audit]
      operationId: listAuditLogs
      summary: Read the audit trail.
      description: |
        Append-only, and written in the same transaction as the change it
        describes: an audit row written in a second transaction is a change that
        can happen unrecorded.

        The cursor is composite (`<RFC 3339 timestamp> <uuid>`) because the table is
        partitioned by `created_at`; an id-only cursor would have to be searched for
        in every partition.
      parameters:
        - $ref: "#/components/parameters/limit"
        - name: cursor
          in: query
          description: An opaque cursor from a previous page's `next_cursor`.
          schema: { type: string }
        - name: action
          in: query
          schema: { type: string, examples: ["deployment.scale"] }
        - name: resource_type
          in: query
          schema: { type: string, examples: ["deployment"] }
        - name: resource_id
          in: query
          schema: { type: string, format: uuid }
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: until
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: A page of audit records.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/AuditLog" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: nbk
      description: |
        `Authorization: Bearer nbk_<key>`. Session tokens share this header in the
        deployed system but are not accepted by this service until the dashboard
        arrives in Phase 7; a non-`nbk_` bearer is refused with code
        `unsupported_credential` rather than a generic 401, because "your
        credential type is not supported yet" and "your key is wrong" are different
        problems.
  parameters:
    limit:
      name: limit
      in: query
      description: Page size, 1-200. Defaults to 50.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    cursor:
      name: cursor
      in: query
      description: |
        The `next_cursor` from a previous page. Cursor-based, not offset-based:
        offset paging skips and duplicates rows while the collection is being
        written to.
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: The request is malformed or semantically invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: |
        No usable credential. The message never distinguishes "no such key" from
        "wrong key": that difference is an oracle for enumerating valid prefixes.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: The credential lacks the required scope or role.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: |
        No such resource. Also returned for a resource in another organization, so
        existence elsewhere is never disclosed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: A uniqueness or state precondition failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      description: |
        One shape everywhere, OpenAI-compatible so existing SDK error handling
        works. It carries nothing internal: no pod names, no SQL detail, no
        upstream bodies. Those belong in logs and traces, which are operator-only.
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message: { type: string }
            type:
              type: string
              enum:
                - invalid_request_error
                - authentication_error
                - permission_error
                - not_found_error
                - conflict_error
                - unprocessable_entity
                - rate_limit_error
                - internal_error
                - upstream_error
                - service_unavailable
                - timeout_error
            code: { type: string }
            param: { type: string }
            request_id: { type: string }
    Page:
      type: object
      required: [data, next_cursor, has_more]
      properties:
        data: { type: array, items: {} }
        next_cursor:
          type: [string, "null"]
          description: Absent when this is the last page.
        has_more: { type: boolean }
    UserRole:
      type: string
      enum: [owner, admin, developer, viewer]
      description: |
        Descending privilege. A role belongs to a human; a scope belongs to a
        credential, and scopes are what the request path checks.
    Scope:
      type: string
      enum:
        - inference:invoke
        - inference:pin
        - models:read
        - models:write
        - deployments:read
        - deployments:write
        - rollouts:write
        - usage:read
        - audit:read
        - admin
      description: |
        `admin` does not imply the others: an admin credential that cannot invoke
        inference is a useful thing to be able to issue.
    Priority:
      type: string
      enum: [LOW, NORMAL, HIGH]
      description: Carried by the credential, never read from a request body.
...
    User:
      type: object
      required: [id, email, role, created_at, updated_at]
      description: The password hash is absent from every serialisation, by construction.
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: [string, "null"] }
        role: { $ref: "#/components/schemas/UserRole" }
        last_login_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateUser:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        name: { type: [string, "null"] }
        role: { $ref: "#/components/schemas/UserRole" }
        password:
          type: string
          minLength: 12
          maxLength: 1024
          description: |
            Optional: a user backed by an external identity provider never holds
            one. Hashed with argon2id. Write-only, never echoed.
    APIKey:
      type: object
      required: [id, name, prefix, scopes, priority, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        prefix:
          type: string
          description: |
            `nbk_` plus 7 characters. Safe to display, log and audit, and the
            indexed column a presented key is looked up by.
        user_id: { type: [string, "null"], format: uuid }
        scopes:
          type: array
          items: { $ref: "#/components/schemas/Scope" }
        priority: { $ref: "#/components/schemas/Priority" }
        last_used_at: { type: [string, "null"], format: date-time }
        expires_at: { type: [string, "null"], format: date-time }
        revoked_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
    CreatedAPIKey:
      allOf:
        - $ref: "#/components/schemas/APIKey"
        - type: object
          required: [key, key_note]
          properties:
            key:
              type: string
              description: The plaintext key. Shown here and never again.
            key_note: { type: string }
    CreateAPIKey:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        user_id:
          type: [string, "null"]
          format: uuid
          description: |
            The human the key belongs to, who must be in the same organization.
            Recorded as `created_by` on resources the key creates.
        scopes:
          type: array
          items: { $ref: "#/components/schemas/Scope" }
          description: |
            Defaults from the creating credential's role rather than to everything,
            so a key created without thought is a limited key. Bounded by the
            creating credential's own scopes.
        priority: { $ref: "#/components/schemas/Priority" }
        expires_at: { type: [string, "null"], format: date-time }
...
    AuditLog:
      type: object
      required: [id, created_at, actor_type, action, resource_type]
      properties:
        id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        actor_type: { type: string, enum: [user, api_key, system] }
        actor_id: { type: [string, "null"], format: uuid }
        actor_label:
          type: [string, "null"]
          description: |
            Denormalised (the key prefix, or the user's email) so the record stays
            readable after the actor is deleted.
        action: { type: string }
        resource_type: { type: string }
        resource_id: { type: [string, "null"], format: uuid }
        before: { type: [object, "null"], additionalProperties: true }
        after: { type: [object, "null"], additionalProperties: true }
        request_id: { type: [string, "null"], format: uuid }
        ip: { type: [string, "null"] }
        user_agent: { type: [string, "null"] }